consololidated preloads into 3.  Adopted the 'sail' namespace for the internal preload api
