+++ packages/preload/src/system/index.ts
@@ -1,0 +1,4 @@
+import { contextBridge } from 'electron';
+import { api } from './api';
+
+contextBridge.exposeInMainWorld('sail', api);
