+++ packages/preload/src/systemView/index.ts
@@ -1,11 +1,11 @@
 import { connect, createAPI } from '../view/api';
 import { contextBridge } from 'electron';
-import { getApps } from './api';
+import { api } from '../system/api';
 connect();
 //listen();
 //document.addEventListener('DOMContentLoaded', createAPI);
 
 /* expose the fdc3 api across the context isolation divide...*/
-contextBridge.exposeInMainWorld('home', { getApps: getApps });
+contextBridge.exposeInMainWorld('sail', api);
 const DesktopAgent = createAPI();
 contextBridge.exposeInMainWorld('fdc3', DesktopAgent);
