+++ packages/preload/vite.config.js
@@ -37,12 +37,9 @@ const config = {
         entryFileNames: '[name].cjs',
       },
       input: {
-        'frame/index': '/src/frame/index.ts',
-        'channelPicker/index': '/src/channelPicker/index.ts',
         'view/index': '/src/view/index.ts',
-        'homeView/index': '/src/homeView/index.ts',
-        'intentResolver/index': '/src/intentResolver/index.ts',
-        'searchResults/index': '/src/searchResults/index.ts',
+        'system/index': '/src/system/index.ts',
+        'systemView/index': '/src/systemView/index.ts',
       },
     },
     emptyOutDir: true,
